% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfStarwars}
\alias{dfStarwars}
\title{Starwars characters}
\format{
A tibble with 87 rows and 11 variables:
\describe{
\item{name}{Name of the character}
\item{height}{Height (cm)}
\item{mass}{Weight (kg)}
\item{hair_color,skin_color,eye_color}{Hair, skin, and eye colors}
\item{birth_year}{Year born (BBY = Before Battle of Yavin)}
\item{sex}{The biological sex of the character, namely male, female, hermaphroditic, or none (as in the case for Droids).}
\item{gender}{The gender role or gender identity of the character as determined by their personality or the way they were programmed (as in the case for Droids).}
\item{homeworld}{Name of homeworld}
\item{species}{Name of species}
}
}
\usage{
dfStarwars
}
\description{
This data comes from SWAPI, the Star Wars API, <https://swapi.dev/>
}
\examples{
dfStarwars
}
\keyword{datasets}
