% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfStarwarsFilms}
\alias{dfStarwarsFilms}
\title{Starwars characters films}
\format{
A tibble with 173 rows and 2 variables:
\describe{
\item{name}{Name of the character}
\item{film}{A film the character appeared in}
}
}
\usage{
dfStarwarsFilms
}
\description{
This data comes from SWAPI, the Star Wars API, <https://swapi.dev/>
}
\examples{
dfStarwarsFilms
}
\keyword{datasets}
