% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfStarwarsStarships}
\alias{dfStarwarsStarships}
\title{Starwars characters starships}
\format{
A tibble with 31 rows and 2 variables:
\describe{
\item{name}{Name of the character}
\item{starship}{A starship the character has piloted}
}
}
\usage{
dfStarwarsStarships
}
\description{
This data comes from SWAPI, the Star Wars API, <https://swapi.dev/>
}
\examples{
dfStarwarsStarships
}
\keyword{datasets}
