% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfStarwarsVehicles}
\alias{dfStarwarsVehicles}
\title{Starwars characters vehicles}
\format{
A tibble with 13 rows and 2 variables:
\describe{
\item{name}{Name of the character}
\item{vehicle}{A vehicle the character has piloted}
}
}
\usage{
dfStarwarsVehicles
}
\description{
This data comes from SWAPI, the Star Wars API, <https://swapi.dev/>
}
\examples{
dfStarwarsVehicles
}
\keyword{datasets}
